some fixes

+++ coroutine.hpp
@@ -389,7 +389,7 @@ public:
     noexcept(noexcept(f()))
     requires(bool(sizeof...(ev)))
   {
-    std::array<int, sizeof...(ev)> r{};
+    std::array<short, sizeof...(ev)> r{};
 
     gnr::forwarder<void(evutil_socket_t, short) noexcept> g(
       [&](evutil_socket_t, short const f) noexcept
